+++ Singleton/Program.cs
@@ -5,7 +5,7 @@ using System.Text;
 
 namespace Singleton
 {
-    class Program
+    static class Program
     {
         static void Main(string[] args)
         {
@@ -25,51 +25,6 @@ namespace Singleton
             Console.WriteLine("Division : " + Calculate.Instance.Division());
 
             Console.ReadLine();
-        }
-    }
-
-    public sealed class Calculate
-    {
-        private Calculate()
-        {
-
-        }
-
-        private static Calculate instance = null;
-        public static Calculate Instance
-        {
-            get
-            {
-                if (instance == null)
-                {
-                    instance = new Calculate();
-                }
-                return instance;
-            }
-        }
-
-
-        public double ValueOne { get; set; }
-        public double ValueTwo { get; set; }
-
-        public double Addition()
-        {
-            return ValueOne + ValueTwo;
-        }
-
-        public double Subtraction()
-        {
-            return ValueOne - ValueTwo;
-        }
-
-        public double Multiplication()
-        {
-            return ValueOne * ValueTwo;
-        }
-
-        public double Division()
-        {
-            return ValueOne / ValueTwo;
         }
     }
 }
